feat: Refactor summary handling and JSON file saving for GitHub Actions
- Consolidated summary writing logic into a dedicated IO module for better organization and reusability.
- Enhanced the main execution flow to initialize and append summaries for issues and projects.
- Improved JSON file saving functionality to handle both issues and projects, ensuring data is saved correctly.
- Removed redundant summary writing code from individual functions, streamlining the overall structure.
