+++ src/index.js
@@ -1,9 +1,18 @@
 //@ts-check
 /// <reference path="./types.d.ts" />
 import * as core from "@actions/core";
-import * as fs from "fs";
+import * as github from "@actions/github";
 import { getAllIssues } from "./issues.js";
 import { getAllProjects } from "./projects.js";
+import {
+  initializeSummary,
+  appendCompletionMessage,
+  appendErrorMessage,
+  appendToSummary,
+  generateIssuesSummaryMarkdown,
+  generateProjectsSummaryMarkdown,
+  saveJsonFiles
+} from "./io.js";
 
 /**
  * メイン実行関数
@@ -15,10 +24,7 @@ async function main() {
   try {
     // GitHub Actions Summaryの初期化
     const summaryPath = process.env.GITHUB_STEP_SUMMARY;
-    if (summaryPath) {
-      // Summaryファイルを初期化（既存の内容をクリア）
-      fs.writeFileSync(summaryPath, '# 📈 GitHub Project Metrics\n\n', 'utf8');
-    }
+    initializeSummary(summaryPath);
     
     // IssueとProjectの両方を取得
     core.info("=== GitHub Project Metrics 実行開始 ===");
@@ -26,24 +32,40 @@ async function main() {
     // Issueを取得
     const issuesData = await getAllIssues();
     
-    // Projectを取得（issuesDataを渡す）
-    await getAllProjects(issuesData);
+    // Projectを取得
+    const projectsData = await getAllProjects();
     
     core.info("=== GitHub Project Metrics 実行完了 ===");
     
-    // Summaryに完了メッセージを追加
+    // GitHub Actions Summaryに書き込む
     if (summaryPath) {
-      fs.appendFileSync(summaryPath, `---\n\n✅ **実行完了**: ${new Date().toLocaleString('ja-JP')}\n`, 'utf8');
+      // IssuesのSummaryを追加
+      const { owner, repo } = github.context.repo;
+      const issuesSummaryMarkdown = generateIssuesSummaryMarkdown(issuesData, owner, repo);
+      appendToSummary(summaryPath, issuesSummaryMarkdown);
+      
+      // ProjectsのSummaryを追加
+      const projectsSummaryMarkdown = generateProjectsSummaryMarkdown(projectsData);
+      appendToSummary(summaryPath, projectsSummaryMarkdown);
     }
+    
+    // JSONファイルを保存
+    try {
+      const outputPath = core.getInput("output-path");
+      saveJsonFiles(outputPath, issuesData, projectsData);
+    } catch (writeError) {
+      core.warning(`Failed to save JSON files: ${writeError.message}`);
+    }
+    
+    // Summaryに完了メッセージを追加
+    appendCompletionMessage(summaryPath);
     
   } catch (error) {
     core.setFailed(error.message);
     
     // エラー時もSummaryに記載
     const summaryPath = process.env.GITHUB_STEP_SUMMARY;
-    if (summaryPath) {
-      fs.appendFileSync(summaryPath, `\n---\n\n❌ **エラー**: ${error.message}\n`, 'utf8');
-    }
+    appendErrorMessage(summaryPath, error.message);
   }
 }
